API Documentation

Docstrings for interface members can be accessed through Julia's built-in documentation system or in the list below.

Contents

API Documentation

Contents

Index

Functions

Index

DocumentationTest.foo

Functions

DocumentationTest.foo — Function

foo(x, y)

Creates a 2-element static array from the scalars x and y.